Everyone gets initalised
