+++ tests/scripts/nif/__init__.py
@@ -24,3 +24,7 @@ def call_niftoaster(*args):
     toaster.cli()
     sys.argv = oldargv
     return toaster
+
+
+class TestNif():
+    pass
