+++ tests/scripts/nif/test_niftoaster.py
